Reorganize code locations
